Bug 2051973 - Vendor and compile the Metal backend for libggml. r?#firefox-build-system-reviewers
This includes the Metal shader source in asm and compiles it. Specific
symbols are then picked up by the implementation, and the byte range is
fed to the Metal shader compiler. The resulting object file is
subsequently cached by the OS.

First compilation is in the order of 4-5s, subsequent instantiation is
10-15ms.

This backend will always ever run in a dedicated process, that has a
sandbox very similar to the GPU process, but doesn't have access to the
display server, fonts, and other facilities related to rendering. It can
access one more folder than the GPU process (read/write) to cache the
compiled shaders.

This benefits llama.cpp, but also anything else that uses ggml such as
parakeet.cpp, that is introduced in a later patch. This can multiply
the performance by multiple dozen times (depending on the hardware
and use-case). More importantly, it completely offload inference
to the GPU. While (e.g.) real-time local speech recognition is
running with this backend, `about:processes` show all thread of
the HWInference process as "idle".

Differential Revision: https://phabricator.services.mozilla.com/D268401