Bug 1940906 - Sandbox policy for the HWInference utility process. r?haik
This is essentially a copy-paste of the GPU process policy, trimmed down
to removed everything that wasn't needed. One thing that has been added
is read/write permissions to the `/plugincontainer` directory so shader
compilation caching works.

We have discussed changing the executable name to relinquish allowing
executable page, this hasn't been implemented.

This also include just enough code so that it compiles and run on Linux,
but that side needs work.

Differential Revision: https://phabricator.services.mozilla.com/D268412