Bug 1940906 - Make the SpeechRecognition side of the Web Speech API mostly up-to-date in terms of spec. r?#webidl-reviwers
https://webaudio.github.io/web-speech-api/#speechreco-section

The SpeechRecognitionBackend object (to be introduced in a subsequent
patch) handles all the "actual" recognition, but this object handles the
DOM side.

A noted in the code, a number of things are unclear in the standard,
that we're working on addressing.

This removes a little bit of the previous backend (online recognition)
that was left after the previous patch removed most of it, and implements:

- available()
- install()
- SpeechRecognitionPhrase
- a new error code
- final attribute of SpeechRecognitionResult