Bug 1940906 - Introduce the HWInference utility process. r?#ipc-reviewers,gerard-majax
The HWInference process is a new type of process, that is created from a
content process (e.g. to implement a Web API), but can also talk to the
main process (e.g. to interact with ModelHub and acquire ML models).

It doesn't run script. It has a dedicated sandbox policy (in a
subsequent patch), that is roughly similar to the GPU process policy,
with everything related to graphics rendering removed (connection to the
display server, access to fonts, misc system calls and privileges,
etc.), as it is just for hardware accelerated computations.

This patch only contains the scaffolding, and the first actor
(PSpeechRecognition) will be introduced in another patch.

Differential Revision: https://phabricator.services.mozilla.com/D268403
Bug 1940906 - DO NOT LAND - Enable SpeechRecognition at build time.

Just for convenience

Differential Revision: https://phabricator.services.mozilla.com/D268418