Bug 1940906 - Implement the backend for downloading a model file, using ModelHub. r?#ipc-reviewers
This is very similar to the previous patch (available()), but leverages
the ModelHub to download a model, with progress callback. The download
progress isn't relayed to content (intentionnaly, for fingerprinting
reasons), but a completion callback is (error or success).

Again, this starts in the content process, uses a transient
PSpeechRecognition actor to reach the HWInference process, that asks the
main process (on its main thread) to download a model (or resolve the
promise immediately if already downloaded -- ModelHub implements a
cache).

Differential Revision: https://phabricator.services.mozilla.com/D268410