Bug 1940906 - Vendor whisper.cpp. r?#firefox-build-system-reviewers,atossou
Whisper.cpp is a C++ speech recognition library that uses models of the
Whisper family. It allows performing local real-time and faster-than
real-time local speech recognition in many languages.

This library uses libggml (that we already vendor in llama.cpp) as its
backend for most computations. I have synchronized the version so it
works well with what we have, and will have to synchronize updates as
well.

This is compiled into `libmozinference`.

A patch is added on top of the upstream library, to be able to load a speech
recognition model from a `FILE*`, instead of a file path. This is passed
down by the main process when starting recognition, similarly to what we
do for llama.cpp. This patch is likely upstreamable.

Differential Revision: https://phabricator.services.mozilla.com/D268402