Bug 1940906 - Introduce PSpeechRecognition and the model-availability IPC path. r?#ipc-reviewers
Introduces the PSpeechRecognition protocol and wires up the availability
round-trip that backs SpeechRecognition.available(): it starts in the
content process, instantiates a SpeechRecognition actor, sends a request
to the HWInference process, which then asks the parent process, at which
point the ModelHub can be queried (on the main thread). The DOM-facing
SpeechRecognition.available() is implemented on top of this in a later
commit.

Differential Revision: https://phabricator.services.mozilla.com/D268409