Bug 1940906 - Make the SpeechRecognition side of the Web Speech API mostly up-to-date in terms of spec. r?#webidl
https://webaudio.github.io/web-speech-api/#speechreco-section

The SpeechRecognitionBackend object introduced here (as a stub) will handle
all the "actual" recognition; this object handles the DOM side. The backend
is fleshed out in subsequent patches.

As noted in the code, a number of things are unclear in the standard, that
we're working on addressing.

This removes a little bit of the previous backend (online recognition) that
was left after the previous patch removed most of it, and implements:

- available()
- install()
- SpeechRecognitionPhrase
- the phrases-not-supported error code (bad-grammar is kept for web
  compatibility even though grammars are not implemented and it never fires)
- the final attribute of SpeechRecognitionResult
- [SecureContext] on the SpeechRecognition interfaces, per spec

most of the WPTs are manual, and need work anyway, this will happen in
another patch as a followup.

Differential Revision: https://phabricator.services.mozilla.com/D268405