Bug 1940906 - Perform speech recognition locally in the browser. r?atossou
This patch receives the audio via IPC, puts it in an SPSC queue. This is
periodically dequeued, and prepared for recognition. The actual
recognition loop is still being tuned and will change, but this is the
general idea.

Two types of result can be sent back: interim result (also called
non-final in the spec), and final results. Interim results are often
early recognition of a complete segment of text. Final results are
usually recognition on a larger part of audio, and it is often the case
that interim results are substring of final results. Sometimes however,
the context brought by new words recognized allow getting a better
recognition in the early part of the sentence. After some time, we
consider the recognition correct, a final result is sent, and the audio
is reset, keeping a little bit, in case we were cutting a word.

Again, this is work in progress, but largely works.