Bug 1940906 - Introduce SpeechRecognitionBackend. r?pehrsons,
This object handles:

- Getting audio from a MediaTrackGraph, downmixing it, queuing it to a ring
buffer
- Running a processing loop on a dedicated thread (called the resampling
thread), that dequeues from the ring buffer, downsamples the audio to 16kHz
(which is the rate at which the inference is running), and sends it over to the
IPC thread, that will send it to the HWInference process in chunks of (by
default) 500ms
- Receive events from the HWInference process, and relay them to the main thread
(notably speech recognition events, but also errors and such) object,
SpeechRecognition
- Running another (static) thread (called the IPC thread) on which the IPC will
run. This is done this way because available and install calls are static, while
the actual speech recognition process require an instance, and our IPC system
requires all IPC calls to be on the same thread.

This also contains some code removal and simplifications for the graph listener.

Differential Revision: https://phabricator.services.mozilla.com/D268407