Bug 1940906 - Vendor mudler/parakeet.cpp for cache-aware streaming speech recognition. r?#firefox-build-system-reviewers
This is mudler/parakeet.cpp, vendored for real-time, cache-aware streaming
speech recognition of a Parakeet-family (RNN-T/joint) model via a streaming
C API (parakeet_capi.h). It uses libggml (that we already vendor in
llama.cpp) as its backend for most computations.

This is compiled into `libmozinference`.

A patch is added on top to be able to load a speech recognition model from a
`FILE*`, instead of a file path. This is passed down by the main process
when starting recognition, similarly to what we do for llama.cpp.

Differential Revision: https://phabricator.services.mozilla.com/D268402