Bug 1940906 - Introduce SpeechRecognitionBackend - Part 1 - Static methods. r?nika
This patch starts filling in the blanks in SpeechRecognitionBackend. This
object sits behind the SpeechRecogntion DOM object, and implements some light
audio processing, and IPC communication.

Audio processing is left for the next patch. This one implements the static
methods, `available` and `install`, that call the IPDL methods of the same name.

More importantly this patch implement a threading architecture that works for
real-time audio, and DOM API, static and instance methods.

The constraints are the following:
- We're going to send method calls (available, install, start, stop, etc.) to
the HWInference process, those come from the main thread (Web Speech API)
- We're also going to need to send real-time audio to the same process (with a
graduingly lower and lower packet size, to reduce end-to-end latency). This come
from a hard-real-time thread (from our MediaTrackGraph)
- Our IPC system requires all method calls to be on a single thread
- The real-time audio thread can't do any system calls or unbounded operations
(can't do IPC, locking, allocations, etc.)
- We really don't want to bounce the real-time audio via the main thread for IPC
or make it more busy if we can avoid it (it's already a nightmare in a typical
media processing web app)
- We can't affort scheduling hazards associated with using a shared thread (e.g.
the shared background thread), because we want to have low and constant latency
- There are both static and instance methods to implement

The design is therefore the following:
- A dedicated SpeechIPC thread is created, and shut down as soon as it's not
needed. It is kept as a static member on the class
- IPC only ever runs on this dedicated thread
- Static method calls are dispatched to that thread, and the result is
dispatched back to the main thread.
- In the next patch, processed and packetized audio is dispatched to that
thread as well
- Whenever a Speech Recognition session is active, or while an available or
install call is in flight, it is kept alive
- Otherwise, we shut it down as soon as we can (session stop, result of
`available` or `install`)

The logic to manage those transient sessions (vs. long running ones) is
encapsulated in a dedicated class.

Differential Revision: https://phabricator.services.mozilla.com/D275653